+++ lib/view/food_add_popup.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:foodadd/model/food_note/food_note_model.dart';
+import 'package:foodadd/view/sharecard.dart';
 
 import 'package:foodadd/model/food_note/food_note_dao.dart';
 import 'package:path_provider/path_provider.dart';
@@ -223,35 +224,5 @@ class _FoodAddDialogState extends State<FoodAddDialog> {
           ],
         )
       );
-  }
-}
-
-class ShareCard extends StatelessWidget {
-  const ShareCard({super.key, required this.food});
-
-  final FoodNote food;
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Colors.blue,
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(food.foodName),
-          Text(food.restName),
-          RatingBarIndicator(
-            rating: food.rating,
-            itemCount: 5,
-            itemSize: 25,
-
-            itemBuilder: (context, _) => const Icon(
-              Icons.star,
-              color: Colors.amber,
-            ),
-          ),
-        ],
-      ),
-    );
   }
 }
